feat: 주물 거절하기

+++ src/main/java/com/sparta/tentenbackend/domain/menu/entity/MenuOption.java
@@ -1,11 +1,21 @@
 package com.sparta.tentenbackend.domain.menu.entity;
 
-import com.sparta.tentenbackend.domain.menu.entity.Menu;
 import com.sparta.tentenbackend.global.BaseEntity;
-import jakarta.persistence.*;
-import java.math.BigInteger;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Table(name = "p_menu_option")
@@ -25,7 +35,7 @@ public class MenuOption extends BaseEntity {
 
     private String category;
 
-    private BigInteger price;
+    private Long price;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "menu_id", nullable = false)
@@ -41,7 +51,6 @@ public class MenuOption extends BaseEntity {
         this.setDeletedAt(java.time.LocalDateTime.now());  // 현재 시간 저장
         this.deletedBy = deletedBy;  // 누가 삭제했는지 저장
     }
-
 
 
     /**
